Add sweet alert

+++ src/pages/dashboard/dataGuru/Record.jsx
@@ -7,14 +7,13 @@ import { useEffect } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import { checkUpdateGuru, createGuruCheck, guruDelete, guruPending, guruRecord, guruSearch, paginateGuru } from '../../../features/dashboard/GuruSlice';
 import { selectAllGuru } from '../../../features/dashboard/GuruSlice';
-import Alert from '../../../components/Alert';
 import ModalCreate from './ModalCreate';
 import ModalUpdate from './ModalUpdate';
-import Message from '../../../components/Message';
 import Loading from '../../../components/Loading';
 import { TabelGurus } from '../../../components/FieldTable';
 import { useCallback } from 'react';
 import Paginate from '../../../components/Paginate';
+import swal from 'sweetalert';
 
 const Record = () => {
     const Menus = AdminMenu;
@@ -24,55 +23,33 @@ const Record = () => {
     const checkCreate = useSelector(createGuruCheck);
     const checkUpdate = useSelector(checkUpdateGuru);
     const [active, setActive] = useState('Data Guru');
-    const [checkAlert, setCheckAlert] = useState(false);
     const [showModal, setShowModal] = useState(false);
     const [showModalUpdate, setShowModalUpdate] = useState(false);
     const [idUpdateModal, setIdUpdateModal] = useState();
-    const [errorData, setErrorData] = useState({
-        message: '',
-        status: '',
-    });
 
     useEffect(() => {
-        if (checkCreate.response) setErrorData({
-            message: 'Data gagal di tambahkan, isi data dengan benar!',
-            status: checkCreate?.response.status
-        });
-        if (checkCreate.message === 'success') setErrorData({
-            message: 'Data berhasil di tambahkan',
-            status: 200,
-        })
-
-    }, [checkCreate]);
-
-    useEffect(() => {
-        if (checkUpdate.response) setErrorData({
-            message: 'Data gagal di edit, isi data dengan benar!',
-            status: checkUpdate?.response.status
-        })
-        if (checkUpdate.message === 'success') setErrorData({
-            message: 'Data berhasil di di edit',
-            status: 200,
-        })
-    }, [checkUpdate]);
-
-    useEffect(() => {
-        dispatch(guruRecord());
-    }, [dispatch]);
+        checkCreate.response && swal("Gagal!", "Data gagal di tambahkan, isi data dengan benar!", "error");
+        checkCreate.message === 'success' && swal('Berhasil!', "Data berhasil di tambahkan", "success");
+        checkUpdate.response && swal("Gagal!", "Data gagal di edit, isi data dengan benar!", "error");
+        checkUpdate.message === 'success' && swal("Berhasil!", "Data berhasil di edit", "success");
+    }, [checkUpdate, checkCreate]);
 
     const handleDelete = useCallback((id) => {
-        dispatch(guruDelete(id));
-        setTimeout(() => {
-            dispatch(guruRecord());
-        }, 500);
+        swal({
+            title: "Yakin ingin hapus data?",
+            icon: "warning",
+            buttons: true,
+            dangerMode: true,
+        }).then((willDelete) => {
+            if (willDelete) {
+                swal("Berhasil!", "Data berhasil di hapus!", "success");
+                dispatch(guruDelete(id));
+                dispatch(guruRecord());
+            }
+        });
+    }, [dispatch]);
 
-        setTimeout(() => {
-            setCheckAlert(true);
-        }, 2000)
-
-        setTimeout(() => {
-            setCheckAlert(false);
-        }, 10000)
+    useEffect(() => {
         dispatch(guruRecord());
     }, [dispatch]);
 
@@ -106,15 +83,6 @@ const Record = () => {
                                 </h1>
                                 <p>Kelola Data Guru</p>
                             </div>
-
-                            {errorData?.status === 422 && (
-                                <Message type={'error'} pesan={errorData.message} />
-                            )}
-
-                            {errorData?.status === 200 && (
-                                <Message type={'success'} pesan={errorData.message} />
-                            )}
-
                             <h1 className='text-lg md:text-xl pb-2 font-medium md:font-semibold md:my-2'>Record Data
                                 <div className="float-right">
                                     <input
@@ -172,9 +140,6 @@ const Record = () => {
                             </div>
                             <Paginate items={dataGuru} dataDispatch={paginateGuru} />
                         </div>
-                    )}
-                    {checkAlert && (
-                        <Alert pesan={'Data berhasil dihapus'} />
                     )}
                 </div>
                 <ModalCreate isVisible={showModal} onClose={() => setShowModal(false)} />
