+++ src/pages/dashboard/dataKelas/Record.jsx
@@ -5,15 +5,14 @@ import { useState } from 'react';
 import { BiTrash, BiEdit } from 'react-icons/bi';
 import ModalCreate from './ModalCreate';
 import ModalUpdate from './ModalUpdate';
-import Alert from '../../../components/Alert';
 import { useDispatch, useSelector } from 'react-redux';
 import { checkCreateKelas, checkUpdateKelas, kelasDelete, kelasRecord, kelasSearch, paginateKelas, pendingKelas, selectAllKelas } from '../../../features/dashboard/KelasSlice';
 import { useEffect } from 'react';
-import Message from '../../../components/Message';
 import Loading from '../../../components/Loading';
 import { TabelKelas } from '../../../components/FieldTable';
 import { useCallback } from 'react';
 import Paginate from '../../../components/Paginate';
+import swal from 'sweetalert';
 
 
 const Record = () => {
@@ -24,54 +23,33 @@ const Record = () => {
     const checkCreate = useSelector(checkCreateKelas);
     const checkUpdate = useSelector(checkUpdateKelas);
     const [active, setActive] = useState('Data Kelas');
-    const [checkAlert, setCheckAlert] = useState(false);
     const [showModalCreate, setShowModalCreate] = useState(false);
     const [showModalUpdate, setShowModalUpdate] = useState(false);
     const [idUser, setIdUser] = useState(null);
-    const [errorData, setErrorData] = useState({
-        message: '',
-        status: '',
-    });
 
     useEffect(() => {
-        if (checkCreate.response) setErrorData({
-            message: 'Data gagal di tambahkan, isi data dengan benar!',
-            status: checkCreate?.response.status
-        });
-        if (checkCreate.message === 'success') setErrorData({
-            message: 'Data berhasil di tambahkan',
-            status: 200,
-        })
-    }, [checkCreate]);
-
-    useEffect(() => {
-        if (checkUpdate.response) setErrorData({
-            message: 'Data gagal di edit, isi data dengan benar!',
-            status: checkUpdate?.response.status
-        })
-        if (checkUpdate.message === 'success') setErrorData({
-            message: 'Data berhasil di di edit',
-            status: 200,
-        })
-    }, [checkUpdate]);
-
-    useEffect(() => {
-        dispatch(kelasRecord());
-    }, [dispatch]);
+        checkCreate.response && swal("Gagal!", "Data gagal di tambahkan, isi data dengan benar!", "error");
+        checkCreate.message === 'success' && swal('Berhasil!', "Data berhasil di tambahkan", "success");
+        checkUpdate.response && swal("Gagal!", "Data gagal di edit, isi data dengan benar!", "error");
+        checkUpdate.message === 'success' && swal("Berhasil!", "Data berhasil di edit", "success");
+    }, [checkCreate, checkUpdate]);
 
     const handleDelete = useCallback((id) => {
-        dispatch(kelasDelete(id));
-        setTimeout(() => {
-            dispatch(kelasRecord());
-        }, 500);
+        swal({
+            title: "Yakin ingin hapus data?",
+            icon: "warning",
+            buttons: true,
+            dangerMode: true,
+        }).then((willDelete) => {
+            if (willDelete) {
+                swal("Berhasil!", "Data berhasil di hapus!", "success");
+                dispatch(kelasDelete(id));
+                dispatch(kelasRecord());
+            }
+        });
+    }, [dispatch]);
 
-        setTimeout(() => {
-            setCheckAlert(true);
-        }, 2000)
-
-        setTimeout(() => {
-            setCheckAlert(false);
-        }, 10000)
+    useEffect(() => {
         dispatch(kelasRecord());
     }, [dispatch]);
 
@@ -104,14 +82,6 @@ const Record = () => {
                                 </h1>
                                 <p>Kelola Data Kelas</p>
                             </div>
-
-                            {errorData?.status === 422 && (
-                                <Message type={'error'} pesan={errorData.message} />
-                            )}
-
-                            {errorData?.status === 200 && (
-                                <Message type={'success'} pesan={errorData.message} />
-                            )}
 
                             <h1 className='text-lg md:text-xl pb-2 font-medium md:font-semibold md:my-2'>Record Data
                                 <div className="float-right">
@@ -164,10 +134,6 @@ const Record = () => {
                             </div>
                             <Paginate items={dataKelas} dataDispatch={paginateKelas} />
                         </div>
-                    )}
-
-                    {checkAlert && (
-                        <Alert pesan={'Data berhasil dihapus'} />
                     )}
                 </div>
                 <ModalCreate isVisible={showModalCreate} onClose={() => setShowModalCreate(false)} />
