+++ src/pages/dashboard/dataMapel/Record.jsx
@@ -8,12 +8,11 @@ import { useDispatch, useSelector } from 'react-redux';
 import { checkCreateMapel, checkUpdateMapel, mapelDelete, mapelRecord, mapelSearch, paginateMapel, pendingMapel, selectAllMapel } from '../../../features/dashboard/MapelSlice';
 import ModalCreate from './ModalCreate';
 import ModalUpdate from './ModalUpdate';
-import Alert from '../../../components/Alert';
-import Message from '../../../components/Message';
 import Loading from '../../../components/Loading';
 import { TabelMapels } from '../../../components/FieldTable';
 import { useCallback } from 'react';
 import Paginate from '../../../components/Paginate';
+import swal from 'sweetalert';
 
 const Record = () => {
     const Menus = AdminMenu;
@@ -25,54 +24,34 @@ const Record = () => {
     const [active, setActive] = useState('Data Mapel');
     const [showModalCreate, setShowModalCreate] = useState(false);
     const [showModalUpdate, setShowModalUpdate] = useState(false);
-    const [checkAlert, setCheckAlert] = useState(false);
     const [idUser, setIdUser] = useState();
-    const [errorData, setErrorData] = useState({
-        message: '',
-        status: '',
-    });
 
     useEffect(() => {
-        if (checkCreate.response) setErrorData({
-            message: 'Data gagal di tambahkan, isi data dengan benar!',
-            status: checkCreate?.response.status
-        });
-        if (checkCreate.message === 'success') setErrorData({
-            message: 'Data berhasil di tambahkan',
-            status: 200,
-        })
-    }, [checkCreate]);
-
-    useEffect(() => {
-        if (checkUpdate.response) setErrorData({
-            message: 'Data gagal di edit, isi data dengan benar!',
-            status: checkUpdate?.response.status
-        })
-        if (checkUpdate.message === 'success') setErrorData({
-            message: 'Data berhasil di di edit',
-            status: 200,
-        })
-    }, [checkUpdate]);
-
-    useEffect(() => {
-        dispatch(mapelRecord());
-    }, [dispatch]);
+        checkCreate.response && swal("Gagal!", "Data gagal di tambahkan, isi data dengan benar!", "error");
+        checkCreate.message === 'success' && swal('Berhasil!', "Data berhasil di tambahkan", "success");
+        checkUpdate.response && swal("Gagal!", "Data gagal di edit, isi data dengan benar!", "error");
+        checkUpdate.message === 'success' && swal("Berhasil!", "Data berhasil di edit", "success");
+    }, [checkCreate, checkUpdate]);
 
     const handleDelete = useCallback((id) => {
-        dispatch(mapelDelete(id));
-        setTimeout(() => {
-            dispatch(mapelRecord());
-        }, 500);
+        swal({
+            title: "Yakin ingin hapus data?",
+            icon: "warning",
+            buttons: true,
+            dangerMode: true,
+        }).then((willDelete) => {
+            if (willDelete) {
+                swal("Berhasil!", "Data berhasil di hapus!", "success");
+                dispatch(mapelDelete(id));
+                dispatch(mapelRecord());
+            }
+        });
+    }, [dispatch]);
 
-        setTimeout(() => {
-            setCheckAlert(true);
-        }, 2000)
-
-        setTimeout(() => {
-            setCheckAlert(false);
-        }, 10000)
+    useEffect(() => {
         dispatch(mapelRecord());
     }, [dispatch]);
+
 
     const handleSearch = useCallback((e) => {
         dispatch(mapelSearch(e.target.value));
@@ -103,15 +82,6 @@ const Record = () => {
                                 </h1>
                                 <p>Kelola Data Mapel</p>
                             </div>
-
-                            {errorData?.status === 422 && (
-                                <Message type={'error'} pesan={errorData.message} />
-                            )}
-
-                            {errorData?.status === 200 && (
-                                <Message type={'success'} pesan={errorData.message} />
-                            )}
-
                             <h1 className='text-lg md:text-xl pb-2 font-medium md:font-semibold md:my-2'>Record Data
                                 <div className="float-right">
                                     <input
@@ -175,9 +145,6 @@ const Record = () => {
                         </div>
                     )}
 
-                    {checkAlert && (
-                        <Alert pesan={'Data berhasil dihapus'} />
-                    )}
                 </div>
                 <ModalCreate isVisible={showModalCreate} onClose={() => setShowModalCreate(false)} />
                 <ModalUpdate isVisible={showModalUpdate} onClose={() => setShowModalUpdate(false)} idUser={idUser} />
